+++ examples/authentication-local-storage/src/app/layout.tsx
@@ -1,26 +1,28 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
+import type { Metadata } from "next"
+import { Geist, Geist_Mono } from "next/font/google"
+import "./globals.css"
+import { AccessTokenResponse, client } from "@epcc-sdk/sdks-shopper"
+import { CREDENTIALS_COOKIE_KEY } from "./constants"
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
   subsets: ["latin"],
-});
+})
 
 const geistMono = Geist_Mono({
   variable: "--font-geist-mono",
   subsets: ["latin"],
-});
+})
 
 export const metadata: Metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
-};
+}
 
 export default function RootLayout({
   children,
 }: Readonly<{
-  children: React.ReactNode;
+  children: React.ReactNode
 }>) {
   return (
     <html lang="en">
@@ -30,5 +32,5 @@ export default function RootLayout({
         {children}
       </body>
     </html>
-  );
+  )
 }
